Added react-router. Implemented the 'library' and 'systemSettings' screens as routes. Reworked the URL design to move away from everything being a query parameter (tab1=x&tab2=y etc...). The Route annotation is now just a String path instead of having tab1,tab2,tab2, and action. Added some input components (textInput, select).

+++ src/main/java/net/ehicks/loon/handlers/admin/BackupHandler.java
@@ -1,9 +1,10 @@
 package net.ehicks.loon.handlers.admin;
 
-import net.ehicks.loon.*;
+import net.ehicks.common.Common;
+import net.ehicks.loon.BackupDbTask;
+import net.ehicks.loon.SystemInfo;
 import net.ehicks.loon.routing.Route;
 import net.ehicks.loon.util.CommonIO;
-import net.ehicks.common.Common;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -11,7 +12,6 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.File;
 import java.io.IOException;
-import java.text.ParseException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
@@ -21,45 +21,44 @@ public class BackupHandler
 {
     private static final Logger log = LoggerFactory.getLogger(BackupHandler.class);
 
-    @Route(tab1 = "admin", tab2 = "backups", tab3 = "", action = "form")
-    public static String showBackups(HttpServletRequest request, HttpServletResponse response) throws ParseException, IOException
+    @Route(path = "admin/backups")
+    public static void showBackups(HttpServletRequest request, HttpServletResponse response) throws IOException
     {
-        UserSession userSession = (UserSession) request.getSession().getAttribute("userSession");
-        File backupDir = new File(SystemInfo.INSTANCE.getBackupDirectory());
-        List<File> backups = new ArrayList<>();
-        if (backupDir.exists() && backupDir.isDirectory())
-            backups = Arrays.asList(backupDir.listFiles());
-        backups.removeIf(file -> !file.getName().contains("loon"));
-        Collections.reverse(backups);
-        request.setAttribute("backups", backups);
+        String action = request.getParameter("action");
+        if (action.equals("form"))
+        {
+            File backupDir = new File(SystemInfo.INSTANCE.getBackupDirectory());
+            List<File> backups = new ArrayList<>();
+            if (backupDir.exists() && backupDir.isDirectory())
+                backups = Arrays.asList(backupDir.listFiles());
+            backups.removeIf(file -> !file.getName().contains("loon"));
+            Collections.reverse(backups);
+            request.setAttribute("backups", backups);
 
-        return "/webroot/admin/backups.jsp";
-    }
+            return;
+        }
 
-    @Route(tab1 = "admin", tab2 = "backups", tab3 = "", action = "create")
-    public static void createBackup(HttpServletRequest request, HttpServletResponse response) throws ParseException, IOException
-    {
-        BackupDbTask.backupToZip();
+        if (action.equals("create"))
+        {
+            BackupDbTask.backupToZip();
 
-        response.sendRedirect("view?tab1=admin&tab2=backups&action=form");
-    }
+            response.sendRedirect("view?tab1=admin&tab2=backups&action=form");
+        }
 
-    @Route(tab1 = "admin", tab2 = "backups", tab3 = "", action = "delete")
-    public static void deleteBackup(HttpServletRequest request, HttpServletResponse response) throws ParseException, IOException
-    {
-        String backupName = Common.getSafeString(request.getParameter("backupName"));
-        File file = new File(SystemInfo.INSTANCE.getBackupDirectory() + backupName);
-        boolean result = file.delete();
-        response.sendRedirect("view?tab1=admin&tab2=backups&action=form");
-    }
+        if (action.equals("delete"))
+        {
+            String backupName = Common.getSafeString(request.getParameter("backupName"));
+            File file = new File(SystemInfo.INSTANCE.getBackupDirectory() + backupName);
+            boolean result = file.delete();
+            response.sendRedirect("view?tab1=admin&tab2=backups&action=form");
+        }
 
-    @Route(tab1 = "admin", tab2 = "backups", tab3 = "", action = "viewBackup")
-    public static void viewBackup(HttpServletRequest request, HttpServletResponse response) throws ParseException, IOException
-    {
-        UserSession userSession = (UserSession) request.getSession().getAttribute("userSession");
-        String backupName = Common.getSafeString(request.getParameter("backupName"));
-        File file = new File(SystemInfo.INSTANCE.getBackupDirectory() + backupName);
+        if (action.equals("viewBackup"))
+        {
+            String backupName = Common.getSafeString(request.getParameter("backupName"));
+            File file = new File(SystemInfo.INSTANCE.getBackupDirectory() + backupName);
 
-        CommonIO.sendFileInResponse(response, file, false);
+            CommonIO.sendFileInResponse(response, file, false);
+        }
     }
 }
